SumoBall: New Enemy and Powerup
New enemy was added to the game, I refactored the original enemy script to to work with the new enemy type and used an override script to change how the enemy functioned for the new enemy while maintaining the base faction.

I also added a new powerup that gives the player the ability to ground pound. I made a prefab variant of the original pick up so it had the same animations and hit box, and changed the mesh render and the tag.

I also refactored the spawn manager to be an array allowing future me to easily add new prefabs into both the enemy and powerup arrays to be spawned.